패키지 구조 동기화
