+++ src/main/java/sideproject/gugumo/api/controller/PostController.java
@@ -1,33 +1,21 @@
-package sideproject.gugumo.controller;
+package sideproject.gugumo.api.controller;
 
-import jakarta.annotation.PostConstruct;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import sideproject.gugumo.cond.MeetingStatusType;
-import sideproject.gugumo.cond.SortType;
-import sideproject.gugumo.domain.Member;
-import sideproject.gugumo.domain.MemberRole;
-import sideproject.gugumo.domain.meeting.GameType;
-import sideproject.gugumo.dto.CustomUserDetails;
-import sideproject.gugumo.dto.detailpostdto.DetailPostDto;
-import sideproject.gugumo.dto.simplepostdto.SimpleTransPostDto;
+import sideproject.gugumo.domain.dto.CustomUserDetails;
+import sideproject.gugumo.domain.dto.detailpostdto.DetailPostDto;
+import sideproject.gugumo.domain.dto.simplepostdto.SimpleTransPostDto;
 import sideproject.gugumo.page.PageCustom;
-import sideproject.gugumo.repository.MemberRepository;
 import sideproject.gugumo.request.CreatePostReq;
 import sideproject.gugumo.request.UpdatePostReq;
 import sideproject.gugumo.response.ApiResponse;
 import sideproject.gugumo.service.PostService;
-import sideproject.gugumo.validate.EnumValue;
 
 
 /**
