+++ src/main/java/sideproject/gugumo/domain/dto/detailpostdto/LongDetailPostDto.java
@@ -1,4 +1,4 @@
-package sideproject.gugumo.dto.simplepostdto;
+package sideproject.gugumo.domain.dto.detailpostdto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,10 +7,9 @@ import lombok.experimental.SuperBuilder;
 import java.time.LocalTime;
 
 @Getter
-@SuperBuilder
 @AllArgsConstructor
-public class SimpleTransLongDto extends SimpleTransPostDto{
-
+@SuperBuilder
+public class LongDetailPostDto extends DetailPostDto{
 
     private LocalTime meetingTime;
     private String meetingDays;
