+++ src/main/java/sideproject/gugumo/domain/dto/simplepostdto/SimpleTransLongDto.java
@@ -1,4 +1,4 @@
-package sideproject.gugumo.dto.detailpostdto;
+package sideproject.gugumo.domain.dto.simplepostdto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,9 +7,10 @@ import lombok.experimental.SuperBuilder;
 import java.time.LocalTime;
 
 @Getter
-@AllArgsConstructor
 @SuperBuilder
-public class LongDetailPostDto extends DetailPostDto{
+@AllArgsConstructor
+public class SimpleTransLongDto extends SimpleTransPostDto{
+
 
     private LocalTime meetingTime;
     private String meetingDays;
