+++ src/main/java/sideproject/gugumo/domain/dto/simplepostdto/SimpleTransShortDto.java
@@ -1,4 +1,4 @@
-package sideproject.gugumo.dto.simplepostdto;
+package sideproject.gugumo.domain.dto.simplepostdto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
