+++ src/main/java/sideproject/gugumo/domain/entity/Member.java
@@ -1,8 +1,7 @@
-package sideproject.gugumo.domain;
+package sideproject.gugumo.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
