+++ src/main/java/sideproject/gugumo/domain/entity/MemberRole.java
@@ -1,4 +1,4 @@
-package sideproject.gugumo.domain;
+package sideproject.gugumo.domain.entity;
 
 public enum MemberRole {
     ROLE_USER, ROLE_MANAGER, ROLE_ADMIN
