+++ src/main/java/sideproject/gugumo/domain/entity/meeting/MeetingStatus.java
@@ -1,0 +1,5 @@
+package sideproject.gugumo.domain.entity.meeting;
+
+public enum MeetingStatus {
+    RECRUIT, END
+}
