+++ src/main/java/sideproject/gugumo/domain/entity/meeting/MeetingType.java
@@ -1,0 +1,5 @@
+package sideproject.gugumo.domain.entity.meeting;
+
+public enum MeetingType {
+    SHORT, LONG
+}
