--- src/main/java/sideproject/gugumo/domain/meeting/MeetingStatus.java
@@ -1,5 +1,0 @@
-package sideproject.gugumo.domain.meeting;
-
-public enum MeetingStatus {
-    RECRUIT, END
-}
