--- src/main/java/sideproject/gugumo/domain/meeting/MeetingType.java
@@ -1,5 +1,0 @@
-package sideproject.gugumo.domain.meeting;
-
-public enum MeetingType {
-    SHORT, LONG
-}
