+++ src/main/java/sideproject/gugumo/filter/JwtFilter.java
@@ -1,4 +1,4 @@
-package sideproject.gugumo.jwt;
+package sideproject.gugumo.filter;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -9,8 +9,9 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
-import sideproject.gugumo.domain.Member;
-import sideproject.gugumo.dto.CustomUserDetails;
+import sideproject.gugumo.domain.entity.Member;
+import sideproject.gugumo.domain.dto.CustomUserDetails;
+import sideproject.gugumo.jwt.JwtUtil;
 
 
 import java.io.IOException;
