+++ src/main/java/sideproject/gugumo/filter/LoginFilter.java
@@ -1,4 +1,4 @@
-package sideproject.gugumo.jwt;
+package sideproject.gugumo.filter;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -10,7 +10,8 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import sideproject.gugumo.dto.CustomUserDetails;
+import sideproject.gugumo.domain.dto.CustomUserDetails;
+import sideproject.gugumo.jwt.JwtUtil;
 
 
 import java.io.IOException;
