+++ src/main/java/sideproject/gugumo/repository/MeetingRepository.java
@@ -2,8 +2,8 @@ package sideproject.gugumo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import sideproject.gugumo.domain.meeting.Meeting;
-import sideproject.gugumo.domain.post.Post;
+import sideproject.gugumo.domain.entity.meeting.Meeting;
+import sideproject.gugumo.domain.entity.post.Post;
 
 import java.util.Optional;
