+++ src/main/java/sideproject/gugumo/repository/MemberRepository.java
@@ -2,7 +2,7 @@ package sideproject.gugumo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import sideproject.gugumo.domain.Member;
+import sideproject.gugumo.domain.entity.Member;
 
 import java.util.Optional;
