+++ src/main/java/sideproject/gugumo/repository/PostRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import sideproject.gugumo.domain.Member;
-import sideproject.gugumo.domain.post.Post;
+import sideproject.gugumo.domain.entity.Member;
+import sideproject.gugumo.domain.entity.post.Post;
 
 import java.util.Optional;
