+++ src/main/java/sideproject/gugumo/repository/PostRepositoryCustom.java
@@ -3,8 +3,8 @@ package sideproject.gugumo.repository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import sideproject.gugumo.cond.PostSearchCondition;
-import sideproject.gugumo.dto.CustomUserDetails;
-import sideproject.gugumo.dto.simplepostdto.SimplePostDto;
+import sideproject.gugumo.domain.dto.CustomUserDetails;
+import sideproject.gugumo.domain.dto.simplepostdto.SimplePostDto;
 
 public interface PostRepositoryCustom {
     Page<SimplePostDto> search(PostSearchCondition cond, Pageable pageable, CustomUserDetails principal);
