+++ src/main/java/sideproject/gugumo/service/PostService.java
@@ -9,17 +9,17 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import sideproject.gugumo.cond.PostSearchCondition;
 import sideproject.gugumo.cond.SortType;
-import sideproject.gugumo.domain.meeting.*;
-import sideproject.gugumo.dto.CustomUserDetails;
-import sideproject.gugumo.dto.detailpostdto.DetailPostDto;
-import sideproject.gugumo.domain.Member;
-import sideproject.gugumo.domain.post.Post;
-import sideproject.gugumo.dto.simplepostdto.SimplePostDto;
-import sideproject.gugumo.dto.detailpostdto.LongDetailPostDto;
-import sideproject.gugumo.dto.detailpostdto.ShortDetailPostDto;
-import sideproject.gugumo.dto.simplepostdto.SimpleTransLongDto;
-import sideproject.gugumo.dto.simplepostdto.SimpleTransPostDto;
-import sideproject.gugumo.dto.simplepostdto.SimpleTransShortDto;
+import sideproject.gugumo.domain.entity.meeting.*;
+import sideproject.gugumo.domain.dto.CustomUserDetails;
+import sideproject.gugumo.domain.dto.detailpostdto.DetailPostDto;
+import sideproject.gugumo.domain.entity.Member;
+import sideproject.gugumo.domain.entity.post.Post;
+import sideproject.gugumo.domain.dto.simplepostdto.SimplePostDto;
+import sideproject.gugumo.domain.dto.detailpostdto.LongDetailPostDto;
+import sideproject.gugumo.domain.dto.detailpostdto.ShortDetailPostDto;
+import sideproject.gugumo.domain.dto.simplepostdto.SimpleTransLongDto;
+import sideproject.gugumo.domain.dto.simplepostdto.SimpleTransPostDto;
+import sideproject.gugumo.domain.dto.simplepostdto.SimpleTransShortDto;
 import sideproject.gugumo.exception.NoAuthorizationException;
 import sideproject.gugumo.exception.exception.PostNotFoundException;
 import sideproject.gugumo.page.PageCustom;
